+++ src/base64ToImageElement.js
@@ -3,10 +3,10 @@
  * @param {string} base64 - URL or base64
  * @return {HTMLImageElement}
  */
-export default function getImageFromBase64 (base64) {
+export default function base64ToImageElement (base64) {
   return new Promise((resolve) => {
     const image = new Image()
-    image.crossOrigin = 'Anonymous' // Avoid CORS error
+    image.crossOrigin = 'anonymous' // Avoid CORS error
     image.src = base64
     image.onload = () => {
       resolve(image)