+++ src/capitalizeFirstLetter.js
@@ -4,12 +4,9 @@
  * @returns {string}
  */
 export default function capitalizeFirstLetter (str) {
-  switch (str.length) {
-    case 0:
-      return ''
-    case 1:
-      return str.charAt(0).toUpperCase()
-    default:
-      return `${str.charAt(0).toUpperCase()}${str.slice(1)}`
+  if (str.length <= 1) {
+    return str.charAt(0).toUpperCase()
+  } else {
+    return `${str.charAt(0).toUpperCase()}${str.slice(1)}`
   }
 }