+++ src/transformImage.js
@@ -1,40 +1,30 @@
-import getFileExtension from './getFileExtension.js'
-
-
 /**
  * Convert an image to base64 string
  * @param {string} src - URL or base64
  * @param {object} options
- * @param {number} [options.extension]
+ * @param {number} options.mime
  * @param {number} [options.width]
  * @param {number} [options.height]
  * @param {number} [options.cropWidth]
  * @param {number} [options.cropHeight]
  * @param {number} [options.quality]
- * @return {string} - base64
+ * @return {Promise<string>} - base64
  */
-export default function imageToBase64 (src, { extension, width, height, cropWidth, cropHeight, quality }) {
+export default function transformImage (src, { mime, width, height, cropWidth, cropHeight, quality }) {
   // Cast options to number
-  width = Number(width) || 0
-  height = Number(height) || 0
-  cropHeight = Number(cropHeight) || 0
-  cropWidth = Number(cropWidth) || 0
-  quality = Number(quality) || 0.8
-  // Extract the file extension from the source
-  extension ??= getFileExtension(src)
+  width = parseInt(width) || 0
+  height = parseInt(height) || 0
+  cropHeight = parseInt(cropHeight) || 0
+  cropWidth = parseInt(cropWidth) || 0
+  quality = parseFloat(quality) || 0.8
 
-  if (!extension) {
-    throw new Error('Cannot retrieve the file extension. You MUST specify the `extension` option.')
+  if (!mime) {
+    throw new Error('The `mime` options is required')
   }
-
-  // jpg is the only exception
-  const mime = extension === 'jpg'
-    ? `image/jpeg`
-    : `image/${extension}`
 
   return new Promise((resolve) => {
     const image = new Image()
-    image.crossOrigin = 'Anonymous' // Avoid CORS error
+    image.crossOrigin = 'anonymous' // Avoid CORS error
     image.src = src
     image.onload = () => {
       // Set image size